+++ app/Users.hs
@@ -27,31 +27,66 @@ import           Control.Monad          (unless)
 
 share [mkPersist sqlSettings, mkMigrate "migrateAll"] [persistLowerCase|
 User
-    token       String
+    token       String           
     servers     (M.Map T.Text Bool) 
     threshold   Int                     default=4
     queued      Bool                    default=false
     lastRenew   UTCTime
-    UniqueToken token
+    UniqueToken token 
     deriving    Show
 |]
 
 initialize db slist = do
     file <- doesFileExist db
-    unless file $ migrateUsers db
-    let defaultServerList = zip (flatten slist) (repeat False)
+    unless file $ do
+        migrateUsers db
+        let defaultServerList = zip (flatten slist) (repeat False)
+        runSqlite (T.pack db) $ insert (User "default" (M.fromList defaultServerList) 0 False (posixSecondsToUTCTime 0))
+        return ()
 
-    runSqlite (T.pack db) $ insert (User "default" (M.fromList defaultServerList) 0 False (posixSecondsToUTCTime 0))
-
-    where flatten ((a,b):xs) = T.pack (a ++ ":" ++ b) : flatten xs 
-          flatten []         = []
+        where flatten ((a,b):xs) = T.pack (a ++ ":" ++ b) : flatten xs 
+              flatten []         = []
 
 migrateUsers db = runSqlite (T.pack db) $ runMigration migrateAll
 
+isQueued db token = runSqlite db $ do
+    user <- selectFirst[UserToken ==. token, UserQueued ==. True][]
+    case user of
+        (Just _) -> return True
+        _        -> return False
+
 addUser db token = runSqlite db $ do
+    user <- selectFirst[UserToken ==. token][]
+    case user of 
+        (Just _) -> return False
+        Nothing  -> do
+            t <- liftIO getCurrentTime
+            (Just (Entity _ def)) <- selectFirst[UserToken ==. "default"][]
+            new <- insert (User token (userServers def) 4 False t)
+            case new of
+                UserKey k -> return True
+                _         -> return False
+    
+addUserWithSList db token slist = runSqlite (T.pack db) $ do
     t <- liftIO getCurrentTime
-    (Just (Entity _ def)) <- selectFirst[UserToken ==. "default"][]
-    insert (User token (userServers def) 4 False t)
+    user <- selectFirst[UserToken ==. token][]
+
+    result <- case user of 
+                (Just (Entity key usr)) -> do
+                    update key [UserServers =. slUpdate slist (userServers usr), UserQueued =. True]
+                    return key
+
+                Nothing -> do
+                    (Just (Entity _ def)) <- selectFirst[UserToken ==. "default"][]
+                    insert (User token (slUpdate slist (userServers def)) 4 True t)
+
+    case result of
+        UserKey k -> return True
+        _         -> return False
+    
+        where slUpdate (s:ss) m = M.update f s (slUpdate ss m)
+              slUpdate [] m     = m
+              f queued          = if queued then Just False else Just True
 
 remUser db token = runSqlite db $ deleteBy (UniqueToken token)
 
@@ -59,9 +94,18 @@ remExpiredUsers db time = do
     t <- getCurrentTime
     let expired = posixSecondsToUTCTime $ utcTimeToPOSIXSeconds t - secondsToNominalDiffTime time
     runSqlite db $ do
-        deleteWhere [UserLastRenew <=. expired]
+        deleteWhere [UserLastRenew <=. expired, UserQueued ==. True]
 
-renewUser db token = runSqlite db $ do
+remUserFromQueue db token = runSqlite (T.pack db) $ do
+    user <- selectFirst[UserToken ==. token, UserQueued ==. True][]
+    case user of 
+        Nothing -> return False
+        (Just (Entity key _)) -> do
+            (Just (Entity _ def)) <- selectFirst[UserToken ==. "default"][]
+            update key [UserQueued =. False, UserServers =. userServers def]
+            return True
+
+renewUser db token = runSqlite (T.pack db) $ do
     user <- selectFirst [UserToken ==. token] []
     case user of 
         (Just (Entity userKey _)) -> do